+++ app/src/main/java/com/gabcode/search_meli/ui/util/ImageUtil.kt
@@ -3,7 +3,7 @@ package com.gabcode.search_meli.ui.util
 import android.util.Log
 import android.widget.ImageView
 import androidx.appcompat.app.AppCompatActivity
-import coil.api.load
+import coil.load
 import com.gabcode.search_meli.ui.extension.imageLoader
 
 object ImageUtil {
@@ -15,7 +15,7 @@ object ImageUtil {
     }
 
     fun loadFull(imageView: ImageView, url: String) {
-        imageView.load(url, imageLoader = (imageView.context as AppCompatActivity).imageLoader){
+        imageView.load(url, imageLoader = (imageView.context as AppCompatActivity).imageLoader) {
             listener(
                 onError = { _ , throwable ->
                     Log.e(TAG, "LoadError", throwable)