+++ CHANGELOG.md
@@ -1,3 +1,12 @@
+# v2.0.0
+## Migrated
+- Migrated Kotlin & Coroutine to v1.6
+- Migrated Gradle to v7.3
+- Migrated SDK min 23 & target 31
+- Migrated Java to v11
+- Migrated Coil to v1.4
+- Migrated OkHttp to v4.9.3 & Dagger to v2.39.1
+
 # v1.0.0
 ## Added
 - Created buildSrc and dependencies